* CRI-validation:
  Add a new assignment to swap01.c
  EffectsWithPointsTo
  Pointers
  Pointers/array16: seems to converge now, no longer a bug
  Pointers/Mensi.sub/exp_intra_etendu05: initialization of file test; Pointers/list00: not part of the current validation, switch to intraprocedural analysis... and chaos followed...
  EffectsWithPointsTo/call36.c: used to debug a problem in call02.c and call22.c
  Pointers/Mensi.sub/exp_intra_etendu05: not a bug in PIPS but in the source code; note: result must be checked by Amira
  Pointers/Mensi.sub/chain: not a bug, but result must be checked by Amira
  Pointers/mensi.sub/tableau_dynamique: seems to work; .bug removed; up to Amira to check the result
  Pointers/Mensi.sub/struct_inter03: not a bug anymore officially, but results are not satisfying I am afraid
  Pointers: updates mostly due to computation of points-to information for condition evaluations; leads to improvements in effects computation since I have reverted Beatrice pessimistic approach in handling translation failures: she generated an anywhere effects, I diagnose an execution failure...
  Pointers: updates mostly due to computation of points-to information for condition evaluations; leads to improvements in effects computation since I have reverted Beatrice pessimistic approach in handling translation failures: she generated an anywhere effects, I diagnose an execution failure...
  Accept changes
  Add inline versions
  Pointers: two updates
  Add inline versions
  Applying an inlining to the same code as if02.c
  Pointers/Mensi.sub/if03: modified version of if02; better script, better returned value, shows lots of problems in the interprocedural analysis, unfortunately
  We need a context-sensitive heap modeling for the interprocedural analysis
  A bugged example: Type mismatch after translation
  Pointers: a few forgotten updates, mostly or all related to error messages
  EffectsWith_ointsTo: updates with fewer anywhere
  Pointers: better aliasing detection at call sites
  Adding a new example to the inerprocedural chapter
  EffectsWithPointTo/PointersArithmetic-future.sub: result and test files added
  Pointers/mensi.sub: additional effect computation added in result for cse_wpt01; cleaned up if01.tpips script and results updated accordingly
  Pointers: improvement in call site handling
  EffectsWithPointsTo: improvements in  call site handling
  Add a new example
  Adding new examples
  Add a new example for the if case
  Add a new example for the exit case
  Add a new example illustrating the case free
  Pointers: some more updates linked to pass name and to arrays of pointers
  Pointers/Strict_typing.sub: improved results for fulguro01 and 02, and list
  Pointers: mostly updates due to the use of the interprocedural analysis as default analysis; the error messages are mostly impacted
  EffectsWithPointsTO: a fe updates, mostly due to the fact that PIPS now forbids the assignment of an undefined value as suggested by Pierre Jouvelot if not the C standard
  Pointers: more updates
  Pointers: updates of results
  PointerValues: I am lost; I assumed I had fixed pv_call04.result/test by having the same number of subscripts in each occurence of *HEAP*_l_13
  Pointers: updates linked to several improvements, mostly interprocedural analysis; lots of pending issues;
  little change of exp_intra_etendu09.c
  Add a new example to illustrate the extended intraprocedural chapter
  Pointers: more updates for interprocedural analysis
  Pointers/Strict_typing.sub: results updated for 'main' functions; no need to have OUT information
  Pointers: results updated for interprocedural points-to analysis
  Pointers: systematic use of interprocedural analysis for points-to analysis
  Pointers: systematic use of interprocedural analysis
  Effects/Old.sub/call19: a new bug linked to changed performed to add a backward translation of formal context effects
  Effects/Effects_With_Pointer_Values.sub: dereferencing02.c has been fixed because the source code was not standard compliant for EffectsWithPointsTo; the source codes are shared by effects with points-to and effects with pointer values; the result is thus changed in this directory too
  Rice: issues linked to effects with points-to
  Rice: issues linked to effects with points-to
  Pointers: bugs, but the failure occurs in effects_with_points_to, in the backward translation of effects at some call sites
  EffectsWithPointsTo: more bugs identified, a few updated test files
  EffectsWithPointsTo: mostly bug files added once the bugs are identified in TODO and updated outputs (test files)
  Pointers/array16: a new convergence bug
  EffectsWithPointsTo: updates and new test cases linked to improvements in backward translation of stub effects
  - Version STEP-3.1
  Pointers: improved interprocedural effects... although they should not be used to validate something in Pointers...
  Pointers/Strict_typing.sub/fabien03: effects are removed, points-to are added in test file
  Pointers/hyantes: several updates
  Pointers/Strict_typing.sub/fabien03: the source code of fabien03.c was bugged
  Pointers/Strict_typing.sub: removal of several .bug files; the new results are not necessarily correct and even less as good as one would expect, but the bug that was specified no longer exists
  Pointers: potentially shared source codes are moved into subdirectories pointer.sub and list.sub
  Pointers: pointer test case improved
  Pointers/mensi.sub: not too sure about the cleanup; should have been validated by Amira Mensi
  Pointers: more forgotten files
  Pointers: more forgotten files
  Pointers: a forgotten later file
  Pointers: a few forgotten tpips files
  DO NOT PUT diff & out files, they are generated!
  Add the points-to analysis to default PIPS validation suite
  Pointers/mensi.sub: updates for a few interprocedural test cases
  Pointers/assignment19: to check a suggestion by Pierre Jouvelot, which is not to propagate undefined values even though this does not create core dumps at execution; points-to analysis still to be improved to handle precisely this case
  Rice: test03 added; the source code is modified by cosntant propagation and dependent types are transformed into constant types, which used to show a bug; stub and heap locations are now removed before a new points-to analysis is started.
  Pointers: updates

+++ packages/PIPS/validation/EffectsWithPointsTo/malloc02.c
@@ -1,4 +1,7 @@
-/* Check type sensitivity. */
+/* Check type sensitivity.
+ *
+ * Bug in second malloc removed because it is not needed for EffectsWithPointsTo
+ */
 
 #include <malloc.h>
 
@@ -8,7 +11,7 @@ int main(int argc, char *argv[])
   float * px;
 
   pi = (int *) malloc(sizeof(int));
-  px = (float *) malloc(sizeof(int));
+  px = (float *) malloc(sizeof(float));
 
   return 0;
 }
